fix: loading indicator
