+++ src/components/Requests/AssignRequest.tsx
@@ -1,35 +1,17 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { Button } from "design-react-kit";
 import { Modal, ModalHeader, ModalBody, ModalFooter } from "reactstrap";
-import { remoteData } from "../../api/common";
 
 const AssignRequest = ({
   assignedToMe,
   original,
-  updateList,
-  setLoading
+  assignAgreements
 }: {
   assignedToMe: boolean;
   original: any;
-  updateList: () => void;
-  setLoading: (l: boolean) => void;
+  assignAgreements(): void;
 }) => {
   const [isOpen, toggleAssign] = useState(false);
-
-  const assignAgreementsMutation =
-    remoteData.Backoffice.Agreement.assignAgreement.useMutation({
-      onSuccess() {
-        updateList();
-      }
-    });
-
-  const assignAgreements = () => {
-    assignAgreementsMutation.mutate({ agreementId: original.id });
-  };
-
-  useEffect(() => {
-    setLoading(assignAgreementsMutation.isLoading);
-  }, [assignAgreementsMutation.isLoading, setLoading]);
 
   const checkAssign = () =>
     original.state === "PendingAgreement"
